Refactor the logic for the transcribing functionality into a dedicated folder
